+++ stream-portfolio/portfolio-http/src/test/java/com/backbase/stream/portfolio/controller/it/WealthPortfolioValuationsIT.java
@@ -9,14 +9,12 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Bean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.web.reactive.function.client.WebClient;
 import com.backbase.stream.portfolio.PortfolioHttpApplication;
 import com.backbase.stream.portfolio.model.ValuationsBundle;
 import com.backbase.stream.portfolio.util.PortfolioHttpTestUtil;
@@ -25,13 +23,13 @@ import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 
 /**
- * WealthPortfolioAllocations IT.
+ * WealthPortfolioValuations IT.
  * 
  * @author Vladimir Kirchev
  *
  */
 @SpringBootTest(classes = PortfolioHttpApplication.class)
-@ContextConfiguration(classes = {WealthPortfolioValuationsIT.TestConfiguration.class})
+@ContextConfiguration(classes = {ItTestConfiguration.class})
 @AutoConfigureWebTestClient(timeout = "20000")
 @ActiveProfiles({"it"})
 class WealthPortfolioValuationsIT {
@@ -43,10 +41,10 @@ class WealthPortfolioValuationsIT {
     @DynamicPropertySource
     static void configureProperties(DynamicPropertyRegistry registry) {
         registry.add("spring.security.oauth2.client.provider.bb.token-uri",
-                () -> String.format("%s/oauth/token", wireMockServer.baseUrl()));
+                     () -> String.format("%s/oauth/token", wireMockServer.baseUrl()));
 
         registry.add("backbase.stream.dbs.portfolio-base-url",
-                () -> String.format("%s/portfolio", wireMockServer.baseUrl()));
+                     () -> String.format("%s/portfolio", wireMockServer.baseUrl()));
     }
 
     @Autowired
@@ -60,8 +58,13 @@ class WealthPortfolioValuationsIT {
         List<ValuationsBundle> valuationsBundles = PortfolioHttpTestUtil.getValuationsBundles();
 
         // When
-        webTestClient.post().uri("/portfolios/valuations/batch").header("Content-Type", "application/json")
-                .header(X_TID_HEADER_NAME, X_TID_HEADER_VALUE).bodyValue(valuationsBundles).exchange().expectStatus()
+        webTestClient.post()
+                .uri("/portfolios/valuations/batch")
+                .header("Content-Type", "application/json")
+                .header(X_TID_HEADER_NAME, X_TID_HEADER_VALUE)
+                .bodyValue(valuationsBundles)
+                .exchange()
+                .expectStatus()
                 .isEqualTo(200);
 
         // Then
@@ -78,61 +81,65 @@ class WealthPortfolioValuationsIT {
     }
 
     private void setupWireMock() {
-        wireMockServer.stubFor(WireMock.post("/oauth/token").willReturn(WireMock.aResponse()
-                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .withBody("{\"access_token\": \"access-token\",\"expires_in\": 600,\"refresh_expires_in\": 1800,"
-                        + "\"refresh_token\": \"refresh-token\",\"token_type\": \"bearer\",\"id_token\": \"id-token\","
-                        + "\"not-before-policy\": 1633622545,"
-                        + "\"session_state\": \"72a28739-3d20-4965-bd86-64410df53d04\",\"scope\": \"openid\"}")));
+        wireMockServer.stubFor(WireMock.post("/oauth/token")
+                .willReturn(WireMock.aResponse()
+                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .withBody("{\"access_token\": \"access-token\",\"expires_in\": 600,\"refresh_expires_in\": 1800,"
+                                + "\"refresh_token\": \"refresh-token\",\"token_type\": \"bearer\",\"id_token\": \"id-token\","
+                                + "\"not-before-policy\": 1633622545,"
+                                + "\"session_state\": \"72a28739-3d20-4965-bd86-64410df53d04\",\"scope\": \"openid\"}")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/1234/valuations?granularity=DAILY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/1234/valuations?granularity=WEEKLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/1234/valuations?granularity=MONTHLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
-        wireMockServer.stubFor(
-                WireMock.delete("/portfolio/integration-api/v1/portfolios/1234/valuations?granularity=QUARTERLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                                .withBody("")));
+        wireMockServer.stubFor(WireMock
+                .delete("/portfolio/integration-api/v1/portfolios/1234/valuations?granularity=QUARTERLY")
+                .willReturn(WireMock.aResponse()
+                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .withBody("")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/5678/valuations?granularity=DAILY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/5678/valuations?granularity=WEEKLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
         wireMockServer
                 .stubFor(WireMock.delete("/portfolio/integration-api/v1/portfolios/5678/valuations?granularity=MONTHLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .willReturn(WireMock.aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                                 .withBody("")));
 
-        wireMockServer.stubFor(
-                WireMock.delete("/portfolio/integration-api/v1/portfolios/5678/valuations?granularity=QUARTERLY")
-                        .willReturn(WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                                .withBody("")));
+        wireMockServer.stubFor(WireMock
+                .delete("/portfolio/integration-api/v1/portfolios/5678/valuations?granularity=QUARTERLY")
+                .willReturn(WireMock.aResponse()
+                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .withBody("")));
 
-        wireMockServer.stubFor(WireMock.put("/portfolio/integration-api/v1/portfolios/1234/valuations").willReturn(
-                WireMock.aResponse().withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE).withBody("")));
-    }
-
-    public static class TestConfiguration {
-        @Bean
-        public WebClient.Builder webClientBuilder() {
-            return WebClient.builder();
-        }
+        wireMockServer.stubFor(WireMock.put("/portfolio/integration-api/v1/portfolios/1234/valuations")
+                .willReturn(WireMock.aResponse()
+                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                        .withBody("")));
     }
 }